knowledge-blog-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic-settings
python-dotenv